requirements.txt
django
django-admin-interface
django-rest-framework
decouple